ml_api/requirements.txt
# Core ML
tensorflow==2.15.0
torch==2.3.1
torchvision==0.18.1

# FastAPI + Server
fastapi==0.115.2
uvicorn==0.34.3

# Image preprocessing & evaluation
scikit-learn==1.4.1
numpy==1.26.4
pillow==10.3.0

python-multipart==0.0.9

aiofiles==23.2.1
starlette==0.46.2
